更改id范围

+++ Back-End/Models/Coupon.cs
@@ -8,7 +8,7 @@ namespace Back_End.Models
     public partial class Coupon
     {
         public int CouponId { get; set; }
-        public byte? CouponTypeId { get; set; }
+        public int? CouponTypeId { get; set; }
         public int? CustomerId { get; set; }
         public DateTime CouponStart { get; set; }
         public DateTime CouponEnd { get; set; }
