+++ Back-End/Models/CustomerGroupCoupon.cs
@@ -8,7 +8,7 @@ namespace Back_End.Models
     public partial class CustomerGroupCoupon
     {
         public byte CustomerLevel { get; set; }
-        public byte CouponTypeId { get; set; }
+        public int CouponTypeId { get; set; }
         public bool CouponNum { get; set; }
 
         public virtual CouponType CouponType { get; set; }
